refactor: extract function to separate packages
